+++ packages/react-behave/src/components/Dropdown/Dropdown.js
@@ -1,44 +1,26 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import { Manager, Popper, Reference, placements } from 'react-popper'
-import ClickOutside from '../ClickOutside'
-import MergeRefs from '../MergeRefs'
-import minWidthModifier from './minWidthModifier'
 import getOtherProps from '../../core/getOtherProps'
-
-function renderContent(renderDropdown, onClickOutside, props) {
-  if (!onClickOutside) {
-    // Don't add the `ClickOutside` component if nobody is listening.
-    return renderDropdown(props)
-  }
-
-  return (
-    <ClickOutside onClickOutside={onClickOutside}>
-      {clickOutsideRef => (
-        <MergeRefs refs={[clickOutsideRef, props.ref]}>
-          {ref => renderDropdown({ ...props, ref })}
-        </MergeRefs>
-      )}
-    </ClickOutside>
-  )
-}
+import {
+  Popper,
+  PopperManager,
+  PopperPlacements,
+  PopperReference,
+} from '../Popper'
 
 /**
- * [create-ref]: https://reactjs.org/docs/react-api.html#reactcreateref
- * [react-popper]: https://github.com/FezVrasta/react-popper
- * [popper-props]: https://github.com/FezVrasta/react-popper#children
+ * [popper-children]: https://github.com/FezVrasta/react-popper#children
  * [popper-placements]: https://popper.js.org/popper-documentation.html#Popper.placements
- * [popper-modifiers]: https://popper.js.org/popper-documentation.html#modifiers
  *
  * Render a dropdown around a component.
  *
  * ## Usage
  *
  * ```jsx
- * import React, { Component } from 'react'
+ * import React from 'react'
  * import { Dropdown } from 'react-behave'
  *
- * class App extends Component {
+ * class App extends React.Component {
  *   state = {
  *     open: false,
  *   }
@@ -47,8 +29,8 @@ function renderContent(renderDropdown, onClickOutside, props) {
  *     this.setState({ open: false })
  *   }
  *
- *   toggleDropDown = () => {
- *     this.setState(({ open }) => ({ open: !open }))
+ *   toggleDropdown = () => {
+ *     this.setState({ open: !this.state.open })
  *   }
  *
  *   render() {
@@ -56,19 +38,17 @@ function renderContent(renderDropdown, onClickOutside, props) {
  *       <Dropdown
  *         onClickOutside={this.handleClickOutside}
  *         open={this.state.open}
- *         renderDropdown={({ open, ref, style }) =>
- *           open && (
- *             <ul ref={ref} style={style}>
- *               <li>Item 1</li>
- *               <li>Item 2</li>
- *               <li>Item 3</li>
- *             </ul>
- *           )
+ *         renderDropdown={({ ref, style }) =>
+ *           <ul ref={ref} style={style}>
+ *             <li>Item 1</li>
+ *             <li>Item 2</li>
+ *             <li>Item 3</li>
+ *           </ul>
  *         }
  *       >
- *         {ref => (
- *           <button ref={ref} onClick={this.toggleDropDown}>
- *             Open drop down
+ *         {({ ref }) => (
+ *           <button ref={ref} onClick={this.toggleDropdown}>
+ *             Open dropdown
  *           </button>
  *         )}
  *       </Dropdown>
@@ -80,49 +60,30 @@ function renderContent(renderDropdown, onClickOutside, props) {
 function Dropdown(props) {
   const otherProps = getOtherProps(Dropdown, props)
   return (
-    <Manager>
-      <Reference>{({ ref }) => props.children(ref)}</Reference>
-      <Popper
-        {...otherProps}
-        placement={props.placement}
-        modifiers={{ minWidthModifier, ...props.modifiers }}
-      >
-        {popperProps =>
-          renderContent(props.renderDropdown, props.onClickOutside, {
-            ...popperProps,
-            open: props.open,
-          })
-        }
-      </Popper>
-    </Manager>
+    <PopperManager>
+      <PopperReference children={props.children} />
+      {props.open && (
+        <Popper
+          {...otherProps}
+          placement={props.placement}
+          children={props.renderDropdown}
+        />
+      )}
+    </PopperManager>
   )
 }
 
 Dropdown.propTypes = {
   /**
-   * _Parameters_: `ref: Object|Function`
+   * _Parameters_: `props: Object`
    *
    * Render the reference component of the dropdown.
-   * `ref` must be passed to the component in order to position correctly the dropdown.
+   *
+   * The `props` object contains:
+   *
+   * - `ref: Object|Function`: Must be passed to the component in order to position correctly the dropdown.
    */
   children: PropTypes.func.isRequired,
-
-  /**
-   * Modifiers used to alter the behavior of your poppers.
-   *
-   * In addition to the default modifiers, `minWidthModifier` is added.
-   * It ensures for top or bottom placements that the width of the dropdown is greater or equal to the width of the reference.
-   *
-   * See [PopperJS's modifiers][popper-modifiers].
-   */
-  modifiers: PropTypes.object,
-
-  /**
-   * _Parameters_: `event: MouseEvent`
-   *
-   * Called for each click outside the dropdown component.
-   */
-  onClickOutside: PropTypes.func,
 
   /**
    * Whether the dropdown should be opened or not.
@@ -134,21 +95,19 @@ Dropdown.propTypes = {
    *
    * Must be one of [PopperJS's placement][popper-placements].
    */
-  placement: PropTypes.oneOf(placements),
+  placement: PropTypes.oneOf(PopperPlacements),
 
   /**
-   * _Parameters_: `ref: Object|Function`, `popperProps: Object`
+   * _Parameters_: `popperProps: Object`
    *
    * Render the dropdown.
-   * `ref` must be passed to the component in order to position correctly the dropdown.
-   * `popperProps` is the object containing styles for positioning that need to be applied on the dropdown component.
-   * This object is the same as the [one provided by react-popper][popper-props] without the `ref`.
+   *
+   * See react-popper's [`Popper.children`][popper-children].
    */
   renderDropdown: PropTypes.func.isRequired,
 }
 
 Dropdown.defaultProps = {
-  modifiers: {},
   open: false,
   placement: 'bottom-start',
 }