+++ packages/react-behave/src/components/Dropdown/Dropdown.test.js
@@ -3,48 +3,52 @@ import Dropdown from './Dropdown'
 import ClickOutside from '../ClickOutside'
 import { mount } from 'enzyme'
 
+function mockRender() {
+  return jest.fn(() => null)
+}
+
 describe('<Dropdown />', () => {
-  it('passes down the refs', () => {
-    const render = jest.fn(() => null)
-    const renderDropdown = jest.fn(() => null)
-    mount(<Dropdown renderDropdown={renderDropdown}>{render}</Dropdown>)
+  it('passes down the ref to `render`', () => {
+    const render = mockRender()
+    mount(<Dropdown renderDropdown={mockRender()}>{render}</Dropdown>)
     expect(render).toHaveBeenCalledWith(expect.anything())
-    expect(renderDropdown).toHaveBeenCalledWith(
-      expect.objectContaining({
-        open: false,
-        ref: expect.anything(),
-        style: expect.any(Object),
-      }),
-    )
   })
 
-  it('should pass down the open state', () => {
-    const renderDropdown = jest.fn(() => null)
+  it('passes down the ref and props to `renderDropdown`', () => {
+    const renderDropdown = mockRender()
     mount(
       <Dropdown open={true} renderDropdown={renderDropdown}>
-        {jest.fn(() => null)}
+        {mockRender()}
       </Dropdown>,
     )
-    expect(renderDropdown).toHaveBeenCalledWith(
-      expect.objectContaining({ open: true }),
-    )
+    expect(renderDropdown).toHaveBeenCalledWith(expect.anything())
+  })
+
+  it("doesn't call `renderDropdown` when not opened", () => {
+    const renderDropdown = mockRender()
+    mount(<Dropdown renderDropdown={renderDropdown}>{mockRender()}</Dropdown>)
+    expect(renderDropdown).not.toHaveBeenCalled()
   })
 
   it("shouldn't add a <ClickOutside /> component if `onClickOutside` is not provided", () => {
-    const renderDropdown = jest.fn(() => null)
+    const renderDropdown = mockRender()
     const wrapper = mount(
-      <Dropdown renderDropdown={renderDropdown}>
-        {jest.fn(() => null)}
+      <Dropdown open={true} renderDropdown={renderDropdown}>
+        {mockRender()}
       </Dropdown>,
     )
     expect(wrapper.find(ClickOutside)).toHaveLength(0)
   })
 
   it('should add a <ClickOutside /> component if `onClickOutside` is provided', () => {
-    const renderDropdown = jest.fn(() => null)
+    const renderDropdown = mockRender()
     const wrapper = mount(
-      <Dropdown onClickOutside={jest.fn()} renderDropdown={renderDropdown}>
-        {jest.fn(() => null)}
+      <Dropdown
+        onClickOutside={mockRender()}
+        open={true}
+        renderDropdown={renderDropdown}
+      >
+        {mockRender()}
       </Dropdown>,
     )
     expect(wrapper.find(ClickOutside).first()).toBeDefined()