+++ packages/react-behave/src/components/Popper/index.js
@@ -1,0 +1,7 @@
+export {
+  Reference as PopperReference,
+  Manager as PopperManager,
+  placements as PopperPlacements,
+} from 'react-popper'
+export { default as Popper } from './Popper'
+export { default as minWidthModifier } from './minWidthModifier'