+++ packages/react-behave/src/components/index.js
@@ -1,5 +1,6 @@
 export { default as ClickOutside } from './ClickOutside'
 export { default as Dropdown } from './Dropdown'
 export { default as MergeRefs } from './MergeRefs'
+export * from './Popper'
 export { default as Responsive } from './Responsive'
 export { default as Select } from './Select'